+++ EOLib/Domain/Login/LoginRequestGrantedData.cs
@@ -9,36 +9,36 @@ namespace EOLib.Domain.Login
     [Record]
     public sealed partial class LoginRequestGrantedData : ITranslatedData
     {
-        public short SessionID { get; }
+        public int SessionID { get; }
         public int CharacterID { get; }
 
-        public short MapID { get; }
+        public int MapID { get; }
         public IReadOnlyList<byte> MapRID { get; }
         public int MapLen { get; }
 
         public int EifRid { get; }
-        public short EifLen { get; }
+        public int EifLen { get; }
         public int EnfRid { get; }
-        public short EnfLen { get; }
+        public int EnfLen { get; }
         public int EsfRid { get; }
-        public short EsfLen { get; }
+        public int EsfLen { get; }
         public int EcfRid { get; }
-        public short EcfLen { get; }
+        public int EcfLen { get; }
 
         public string Name { get; }
         public string Title { get; }
         public string GuildName { get; }
         public string GuildRank { get; }
-        public byte ClassID { get; }
+        public int ClassID { get; }
         public string GuildTag { get; }
         public AdminLevel AdminLevel { get; }
 
         public CharacterStats CharacterStats { get; }
 
-        public IReadOnlyDictionary<EquipLocation, short> Paperdoll { get; }
+        public IReadOnlyDictionary<EquipLocation, int> Paperdoll { get; }
 
-        public byte GuildRankNum { get; }
-        public short JailMap { get; }
+        public int GuildRankNum { get; }
+        public int JailMap { get; }
         public bool FirstTimePlayer { get; }
     }
 }
